requirements.txt
binance-futures-connector==3.2.0
python-dotenv==0.21.0
websocket-client==1.2.0
python-telegram-bot==20.1
